refactor: useWalletService
