+++ packages/mask/src/plugins/WalletService/database/Plugin.db.ts
@@ -1,6 +1,6 @@
+import type { Plugin } from '@masknet/plugin-infra'
 import type { NetworkRecord } from '../services/wallet/database/network.js'
 import type { SecretRecord, WalletRecord } from '../services/wallet/type.js'
-import type { Plugin } from '@masknet/plugin-infra'
 
 export let PluginDB: Plugin.Worker.DatabaseStorage<WalletRecord | SecretRecord | NetworkRecord>
